feat: exclude `_timescaledb_internal` tables from structure.sql.
`_compressed_hypertable_*` tables were causing issues when enabling
 compression on hypertables.
