+++ gemfiles/7.0.gemfile.lock
@@ -1,7 +1,7 @@
 PATH
   remote: ..
   specs:
-    timescaledb-rails (0.0.1)
+    timescaledb-rails (0.1.1)
       rails (>= 6.0)
 
 GEM
