+++ lib/timescaledb/rails/railtie.rb
@@ -3,6 +3,7 @@
 require 'rails'
 
 require_relative 'extensions/active_record/database_tasks'
+require_relative 'extensions/active_record/postgresql_database_tasks'
 require_relative 'extensions/active_record/schema_dumper'
 require_relative 'extensions/active_record/schema_statements'
 
@@ -18,8 +19,12 @@ module Timescaledb
 
       initializer 'timescaledb-rails.add_timescale_support_to_active_record' do
         ActiveSupport.on_load(:active_record) do
-          ::ActiveRecord::Tasks::PostgreSQLDatabaseTasks.prepend(
+          ::ActiveRecord::Tasks::DatabaseTasks.extend(
             Timescaledb::Rails::ActiveRecord::DatabaseTasks
+          )
+
+          ::ActiveRecord::Tasks::PostgreSQLDatabaseTasks.prepend(
+            Timescaledb::Rails::ActiveRecord::PostgreSQLDatabaseTasks
           )
 
           ::ActiveRecord::ConnectionAdapters::PostgreSQLAdapter.prepend(
